web_app/scrape_app/src/requirements.txt
mysql-connector-python
selenium
dotenv
pandas